refactor(engine): extract include lowering into lower/include.rs
The deferred-mask + include-splice walk grew to a handful of mutually
recursive helpers; lift them into their own module under
`crates/toasty/src/engine/lower/include.rs` alongside `relation.rs`,
`returning.rs`, and friends. The module owns:

- `process_top_level_includes` (entry from `visit_returning_mut`)
- `walk_record_fields` and `walk_enum_arms` (mapping-driven traversal)
- `dispatch_slot` and `descend_into_embed`
- `build_include_subquery` (relation case)
- two free helpers (`partition_paths`, `loaded_form`)

Renames `process_includes`/`process_field_slot`/`recurse_embed`/
`process_enum_arms` to `walk_record_fields`/`dispatch_slot`/
`descend_into_embed`/`walk_enum_arms` for clarity. Module-level docs
diagram the call graph. Also drops the unused `_sub_paths` parameter
from the enum walker.

No behavior change. `lower.rs` shrinks by 341 lines.
