+++ crates/toasty/src/engine/lower.rs
@@ -1,4 +1,5 @@
 mod expr_pattern;
+mod include;
 mod insert;
 mod paginate;
 mod relation;
@@ -13,7 +14,7 @@ use toasty_core::{
     Result, Schema,
     driver::Capability,
     schema::{
-        app::{self, FieldTy, ModelRoot},
+        app::{self, ModelRoot},
         db::ColumnId,
         mapping,
     },
@@ -836,346 +837,6 @@ impl<'a, 'b> LowerStatement<'a, 'b> {
             }
             _ => todo!("cx={:#?}", self.cx),
         }
-    }
-
-    /// Top-level entry: dispatches every model field on `mapping.default_returning`.
-    /// Reads the root model's app fields and field mappings via
-    /// `self.schema()` (long-lived `'b` borrow) so the recursive helper can
-    /// freely re-borrow `&mut self` for relation include subqueries.
-    fn process_top_level_includes(
-        &mut self,
-        returning: &mut stmt::Expr,
-        model_id: app::ModelId,
-        include_paths: &[stmt::Projection],
-        is_insert: bool,
-    ) {
-        let app_fields: &[app::Field] = self
-            .schema()
-            .app
-            .model(model_id)
-            .as_root_unwrap()
-            .fields
-            .as_slice();
-        let mapping_fields: &[mapping::Field] =
-            self.schema().mapping_for(model_id).fields.as_slice();
-        self.process_includes(
-            returning,
-            app_fields,
-            mapping_fields,
-            include_paths,
-            is_insert,
-        );
-    }
-
-    /// Walks the schema mapping tree (`app_fields`/`mapping_fields`) in
-    /// lockstep with the `returning` record and splices loaded forms at slots
-    /// named by `include_paths` — or, when `is_insert`, at every deferred
-    /// slot at every depth.
-    ///
-    /// The recursion is mapping-driven: each `mapping::Field` variant decides
-    /// how to descend into the corresponding slot expression. Driving off the
-    /// mapping (rather than pattern-matching the slot expression) is what
-    /// lets us reach a `#[deferred]` sub-field of an embed struct nested
-    /// inside an enum variant — those slots live inside a `Match`
-    /// expression, not a `Record`.
-    fn process_includes(
-        &mut self,
-        returning: &mut stmt::Expr,
-        app_fields: &[app::Field],
-        mapping_fields: &[mapping::Field],
-        include_paths: &[stmt::Projection],
-        is_insert: bool,
-    ) {
-        let stmt::Expr::Record(record) = returning else {
-            return;
-        };
-
-        for (i, (field, mapping)) in app_fields.iter().zip(mapping_fields).enumerate() {
-            // Partition include paths matching this slot. `bare` covers
-            // `[i]` ("include this field"); `sub_paths` covers `[i, …]`
-            // ("include sub-path within"). Either kind activates the slot.
-            let mut bare = false;
-            let mut sub_paths: Vec<stmt::Projection> = Vec::new();
-            for path in include_paths {
-                if let Some((first, rest)) = path.as_slice().split_first()
-                    && *first == i
-                {
-                    if rest.is_empty() {
-                        bare = true;
-                    } else {
-                        sub_paths.push(stmt::Projection::from(rest));
-                    }
-                }
-            }
-
-            // Relations are spliced as a subquery into the parent record.
-            // They never live inside an embedded type today, so we handle
-            // them here where `record` is in scope rather than threading the
-            // parent through every recursive helper.
-            if field.ty.is_relation() {
-                if bare || !sub_paths.is_empty() {
-                    self.build_include_subquery(record, i, &sub_paths);
-                }
-                continue;
-            }
-
-            self.process_field_slot(&mut record[i], field, mapping, bare, &sub_paths, is_insert);
-        }
-    }
-
-    /// Dispatches one non-relation field slot by `mapping::Field` kind.
-    ///
-    /// - **Deferred field** — when activated (by include match or `is_insert`),
-    ///   wraps the slot in `Record([loaded])` (column ref for primitives, the
-    ///   embed's `default_returning` for embed targets) and recurses into
-    ///   the wrapped inner so nested deferred sub-fields apply.
-    /// - **Eager embed (struct)** — recurses into the slot's record.
-    /// - **Eager embed (enum)** — recurses into each `Match` arm so
-    ///   `#[deferred]` sub-fields living inside an embed-struct variant
-    ///   field still get spliced.
-    /// - **Eager primitive** — no-op.
-    fn process_field_slot(
-        &mut self,
-        slot: &mut stmt::Expr,
-        field: &app::Field,
-        mapping: &mapping::Field,
-        bare: bool,
-        sub_paths: &[stmt::Projection],
-        is_insert: bool,
-    ) {
-        if field.deferred {
-            let loaded = is_insert || bare || !sub_paths.is_empty();
-            if !loaded {
-                return;
-            }
-
-            let loaded_inner = match (&field.ty, mapping) {
-                (app::FieldTy::Primitive(_), mapping::Field::Primitive(p)) => p.column_expr.clone(),
-                (app::FieldTy::Embedded(_), mapping::Field::Struct(s)) => {
-                    s.default_returning.clone()
-                }
-                (app::FieldTy::Embedded(_), mapping::Field::Enum(e)) => e.default_returning.clone(),
-                _ => unreachable!("deferred field has unexpected mapping shape"),
-            };
-            *slot = stmt::Expr::record([loaded_inner]);
-
-            // Descend into the wrapped inner so nested deferred sub-fields
-            // get spliced if the caller named them or if `is_insert`.
-            if let app::FieldTy::Embedded(embedded) = &field.ty {
-                let stmt::Expr::Record(outer) = slot else {
-                    unreachable!("just-wrapped slot");
-                };
-                self.recurse_embed(
-                    &mut outer[0],
-                    embedded.target,
-                    mapping,
-                    sub_paths,
-                    is_insert,
-                );
-            }
-            return;
-        }
-
-        // Eager field — only embeds need recursion; eager primitives sit on
-        // the column reference already in `default_returning`.
-        if let app::FieldTy::Embedded(embedded) = &field.ty
-            && (is_insert || !sub_paths.is_empty())
-        {
-            self.recurse_embed(slot, embedded.target, mapping, sub_paths, is_insert);
-        }
-    }
-
-    /// Descends into an embed slot. For struct embeds the slot is a `Record`;
-    /// for enum embeds it is a `Match` (or a bare column ref for unit-only
-    /// enums — nothing to descend into).
-    fn recurse_embed(
-        &mut self,
-        slot: &mut stmt::Expr,
-        target: app::ModelId,
-        mapping: &mapping::Field,
-        sub_paths: &[stmt::Projection],
-        is_insert: bool,
-    ) {
-        match (self.schema().app.model(target), mapping) {
-            (app::Model::EmbeddedStruct(em), mapping::Field::Struct(fs)) => {
-                self.process_includes(
-                    slot,
-                    em.fields.as_slice(),
-                    fs.fields.as_slice(),
-                    sub_paths,
-                    is_insert,
-                );
-            }
-            (app::Model::EmbeddedEnum(em), mapping::Field::Enum(fe)) => {
-                self.process_enum_arms(slot, em, fe, is_insert);
-            }
-            _ => {}
-        }
-    }
-
-    /// Walks each arm of an embedded enum's `Match` expression and dispatches
-    /// the variant fields the same way as a struct's record fields.
-    ///
-    /// Each data-arm record has the discriminant at position 0 and variant
-    /// fields at positions 1.., so iterating variant fields offsets the arm
-    /// record index by one.
-    ///
-    /// Include paths into specific variants don't exist today (the path
-    /// macro has no syntax for it), so this only services the `is_insert`
-    /// short-circuit. The shape is what lets a deferred sub-field nested
-    /// inside a variant's embed-struct round-trip through `INSERT…RETURNING`.
-    fn process_enum_arms(
-        &mut self,
-        slot: &mut stmt::Expr,
-        app_enum: &app::EmbeddedEnum,
-        mapping: &mapping::FieldEnum,
-        is_insert: bool,
-    ) {
-        let stmt::Expr::Match(match_expr) = slot else {
-            return;
-        };
-
-        for (variant_idx, arm) in match_expr.arms.iter_mut().enumerate() {
-            let variant_fields: Vec<&app::Field> = app_enum.variant_fields(variant_idx).collect();
-            if variant_fields.is_empty() {
-                continue;
-            }
-            let stmt::Expr::Record(arm_record) = &mut arm.expr else {
-                continue;
-            };
-            let variant_mapping = &mapping.variants[variant_idx];
-
-            for (j, (var_field, var_mapping)) in variant_fields
-                .iter()
-                .zip(&variant_mapping.fields)
-                .enumerate()
-            {
-                self.process_field_slot(
-                    &mut arm_record[j + 1],
-                    var_field,
-                    var_mapping,
-                    false,
-                    &[],
-                    is_insert,
-                );
-            }
-        }
-        // The else arm holds `Expr::Error` placeholders for unreachable
-        // variants; nothing to splice there.
-    }
-
-    /// Builds the relation subquery to splice into `returning[field_index]`
-    /// for a `.include()` of a `BelongsTo`/`HasMany`/`HasOne`. Called only by
-    /// `process_includes` for relation slots; deferred and embedded slots
-    /// never reach here.
-    fn build_include_subquery(
-        &mut self,
-        returning: &mut stmt::ExprRecord,
-        field_index: usize,
-        nested: &[stmt::Projection],
-    ) {
-        let field = &self.model_unwrap().fields[field_index];
-
-        let (mut stmt, target_model_id) = match &field.ty {
-            FieldTy::HasMany(rel) => (
-                stmt::Query::new_select(
-                    rel.target,
-                    stmt::Expr::eq(
-                        stmt::Expr::ref_parent_model(),
-                        stmt::Expr::ref_self_field(rel.pair),
-                    ),
-                ),
-                rel.target,
-            ),
-            // To handle single relations, we need a new query modifier that
-            // returns a single record and not a list. This matters for the type
-            // system.
-            FieldTy::BelongsTo(rel) => {
-                let source_fk;
-                let target_pk;
-
-                if let [fk_field] = &rel.foreign_key.fields[..] {
-                    source_fk = stmt::Expr::ref_parent_field(fk_field.source);
-                    target_pk = stmt::Expr::ref_self_field(fk_field.target);
-                } else {
-                    let mut source_fk_fields = vec![];
-                    let mut target_pk_fields = vec![];
-
-                    for fk_field in &rel.foreign_key.fields {
-                        source_fk_fields.push(stmt::Expr::ref_parent_field(fk_field.source));
-                        target_pk_fields.push(stmt::Expr::ref_parent_field(fk_field.source));
-                    }
-
-                    source_fk = stmt::Expr::record_from_vec(source_fk_fields);
-                    target_pk = stmt::Expr::record_from_vec(target_pk_fields);
-                }
-
-                let mut query =
-                    stmt::Query::new_select(rel.target, stmt::Expr::eq(source_fk, target_pk));
-                query.single = true;
-                (query, rel.target)
-            }
-            FieldTy::HasOne(rel) => {
-                let mut query = stmt::Query::new_select(
-                    rel.target,
-                    stmt::Expr::eq(
-                        stmt::Expr::ref_parent_model(),
-                        stmt::Expr::ref_self_field(rel.pair),
-                    ),
-                );
-                query.single = true;
-                (query, rel.target)
-            }
-            _ => todo!(),
-        };
-
-        // Attach each non-empty remainder as a nested include on the subquery.
-        // Empty remainders (from a bare `.include(posts())`) need no nested
-        // include — the subquery itself satisfies them. The lowering pipeline
-        // will recursively group and process the nested includes when it
-        // encounters `Returning::Model` on this subquery.
-        for rest in nested {
-            if !rest.is_empty() {
-                stmt.include(stmt::Path {
-                    root: stmt::PathRoot::Model(target_model_id),
-                    projection: rest.clone(),
-                });
-            }
-        }
-
-        // Simplify the new stmt to handle relations.
-        Simplify::with_context(self.expr_cx).visit_stmt_query_mut(&mut stmt);
-
-        let mut sub_expr = stmt::Expr::stmt(stmt);
-
-        // For nullable single relations (HasOne<Option<T>>, BelongsTo<Option<T>>),
-        // wrap the sub-expression with a Let + Match to encode the result using
-        // variant-encoded values that distinguish loaded-None from unloaded.
-        //
-        //   Let {
-        //     binding: Stmt(query),
-        //     body: Match {
-        //       subject: Arg(0),
-        //       arms: [Null → I64(0)],
-        //       else_: Arg(0)
-        //     }
-        //   }
-        if field.nullable() && !field.ty.is_has_many() {
-            sub_expr = stmt::Expr::Let(stmt::ExprLet {
-                bindings: vec![sub_expr],
-                body: Box::new(stmt::Expr::match_expr(
-                    stmt::Expr::arg(0),
-                    vec![stmt::MatchArm {
-                        pattern: stmt::Value::Null,
-                        expr: stmt::Expr::from(0i64),
-                    }],
-                    // Non-null: pass through as-is (raw model record)
-                    stmt::Expr::arg(0),
-                )),
-            });
-        }
-
-        returning[field_index] = sub_expr;
     }
 
     /// Returns the ArgId for the new reference
